refactor(cli): split envelope, hints, schema, output into internal modules
Split envelope.ts (694→323), hints.ts (591→98), schema.ts (563→47),
output.ts (416→199). Add eslint-disable to command.ts (fluent builder)
and colors.ts (nested dir breaks bunup shared chunk paths).

🤘🏻 In-collaboration-with: [Claude Code](https://claude.com/claude-code)
